chore: move shouldSkipFetchDueToFetchFailures to utils since its pure
